requirements-dev.txt
# Development dependencies
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.0.0
factory-boy==3.3.0
django-debug-toolbar==4.3.0